requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
supabase
pgvector
agno
httpx
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
googlemaps
python-json-logger
pandas
google-genai
aiofiles
langchain-postgres
duckduckgo-search
openpyxl
langchain-google-genai
jinja2
Pillow
pytest
httpx
pytest-asyncio
